Add tools.py with Line and GpsManager classes.

+++ calibration.py
@@ -34,6 +34,6 @@ if __name__ == "__main__":
                 for k in range(n):
                     measurements[:, k] = np.array(imu.read_mag_raw())
                     time.sleep(0.01)
-                res[i] = np.median(measurements, axis=1, keepdims=True).astype(int)
+                res[i] = np.median(measurements, axis=1, keepdims=True)
                 print("%s = (%f, %f, %f)" % (label, *list(res[i].flatten())))
             np.save(f, res)
